deploy(agentcore): the forwarder is a file, not a template literal
A 230-line Lambda program lived inside plan.ts as an escaped template
string — unlinted, unformatted, with \` and \${ throughout — because one
value (MAX_WEBHOOK_BODY_BYTES) was interpolated into it. That value now
rides in as a Lambda environment variable the template sets, and
forwarder.js is a real CommonJS file beside its reader: biome checks it,
the postbuild copies it into dist, and forwarderSource() reads it.
