+++ knip.jsonc
@@ -18,7 +18,14 @@
     "test/**/*.test.ts"
   ],
   "project": ["src/**/*.ts", "test/**/*.ts"],
-  // Scaffold bundles/templates are copied as files at scaffold time, never imported; fixtures are test data.
-  "ignore": ["src/channels/*/scaffold/**", "src/scaffold/templates/**", "test/fixtures/**"],
+  // Scaffold bundles/templates are copied as files at scaffold time, never imported; fixtures are test
+  // data; the AgentCore forwarder is read as text and runs on Lambda, where its @aws-sdk requires are
+  // the runtime's.
+  "ignore": [
+    "src/channels/*/scaffold/**",
+    "src/scaffold/templates/**",
+    "src/deploy/agentcore/forwarder.js",
+    "test/fixtures/**"
+  ],
   "ignoreBinaries": ["fastagent", "cloudflared"]
 }
