+++ src/deploy/agentcore/forwarder.js
@@ -1,0 +1,244 @@
+// Generated by `fastagent deploy agentcore` — the deployment's only ingress.
+// REGENERATED ON EVERY DEPLOY; edits here are overwritten and never deployed. `--run` builds the
+// Lambda package from fastagent's own copy of this source, so this file is the readable MIRROR of
+// what runs (and what the manual runbook zips) — not an input you can change.
+// Webhooks (Function URL) and EventBridge Scheduler fires are forwarded as envelopes to the
+// AgentCore Runtime over SigV4 InvokeAgentRuntime, all on ONE fixed ingress session (fastagent
+// channel state is single-writer; one session = at most one microVM). With selfSchedule, this
+// Lambda also OWNS the wake alarms: the container POSTs its pending wake-ups to /__fastagent/
+// wake-alarm (shared secret) and each becomes a self-deleting one-shot EventBridge schedule that
+// pokes this Lambda — which wakes the container, whose wake pump fires the due entry.
+// CommonJS on purpose: the deployment package's entry lands as index.js, where ESM import is invalid.
+"use strict";
+const crypto = require("node:crypto");
+const { BedrockAgentCoreClient, InvokeAgentRuntimeCommand } = require("@aws-sdk/client-bedrock-agentcore");
+const client = new BedrockAgentCoreClient({});
+let ownUrl; // self-resolved once per cold start; rides on every envelope for the wake-alarm callback
+
+// Presigned S3 URLs for the container's state snapshot. AgentCore wipes the /mnt/state mount on
+// every runtime version update (= every deploy), so the durable copy lives in S3 — but the
+// container is given NO AWS credentials by the platform, so the only reachable form is a URL that
+// carries its own authorization. SigV4 query signing, node:crypto only (no SDK, nothing to install).
+const enc = (s) => encodeURIComponent(s).replace(/[!'()*]/g, (c) => `%${c.charCodeAt(0).toString(16).toUpperCase()}`);
+const hmac = (key, data) => crypto.createHmac("sha256", key).update(data).digest();
+
+// Every shared-secret gate on this public URL goes through here. Constant-time, and an unset
+// expected secret NEVER matches — otherwise a topology that did not configure one would accept a
+// request that sent none. Non-string input is coerced to "" first: Buffer.from(8) allocates eight
+// bytes rather than throwing, which would compare a number against a secret by length alone.
+const secretEq = (given, expected) => {
+  const a = Buffer.from(typeof given === "string" ? given : ""),
+    b = Buffer.from(expected || "");
+  return b.length > 0 && a.length === b.length && crypto.timingSafeEqual(a, b);
+};
+
+function presign(method, seconds) {
+  const bucket = process.env.STATE_BUCKET,
+    key = process.env.STATE_KEY,
+    region = process.env.AWS_REGION;
+  const host = `${bucket}.s3.${region}.amazonaws.com`;
+  const stamp = new Date().toISOString().replace(/[-:]/g, "").replace(/\.\d+/, "");
+  const scope = `${stamp.slice(0, 8)}/${region}/s3/aws4_request`;
+  const pairs = [
+    ["X-Amz-Algorithm", "AWS4-HMAC-SHA256"],
+    ["X-Amz-Credential", `${process.env.AWS_ACCESS_KEY_ID}/${scope}`],
+    ["X-Amz-Date", stamp],
+    ["X-Amz-Expires", String(seconds)],
+    ["X-Amz-SignedHeaders", "host"],
+  ];
+  if (process.env.AWS_SESSION_TOKEN) pairs.push(["X-Amz-Security-Token", process.env.AWS_SESSION_TOKEN]);
+  // The canonical query must be byte-identical to the one on the wire — build it ONCE, reuse below.
+  const query = pairs
+    .map(([k, v]) => [enc(k), enc(v)])
+    .sort((a, b) => (a[0] < b[0] ? -1 : 1))
+    .map((p) => p.join("="))
+    .join("&");
+  const path = `/${key.split("/").map(enc).join("/")}`;
+  const canonical = [method, path, query, `host:${host}\n`, "host", "UNSIGNED-PAYLOAD"].join("\n");
+  const sts = ["AWS4-HMAC-SHA256", stamp, scope, crypto.createHash("sha256").update(canonical).digest("hex")].join(
+    "\n",
+  );
+  let k = hmac(`AWS4${process.env.AWS_SECRET_ACCESS_KEY}`, stamp.slice(0, 8));
+  for (const part of [region, "s3", "aws4_request"]) k = hmac(k, part);
+  return `https://${host}${path}?${query}&X-Amz-Signature=${hmac(k, sts).toString("hex")}`;
+}
+
+async function invoke(envelope) {
+  if ((process.env.WAKE_SECRET || process.env.STATE_REFRESH_SECRET) && !ownUrl) {
+    const { LambdaClient, GetFunctionUrlConfigCommand } = require("@aws-sdk/client-lambda");
+    ownUrl = (
+      await new LambdaClient({}).send(
+        new GetFunctionUrlConfigCommand({ FunctionName: process.env.AWS_LAMBDA_FUNCTION_NAME }),
+      )
+    ).FunctionUrl;
+  }
+  if (ownUrl) envelope.wake = { url: ownUrl };
+  // Authenticates this envelope as coming from the forwarder (see the template's FastagentIngressSecret).
+  envelope.auth = process.env.INGRESS_SECRET;
+  // Keep each capability short-lived. Function-URL deployments also carry an authenticated refresh
+  // endpoint, so a background turn settling hours after its webhook never depends on the temporary
+  // Lambda credentials that signed the original pair still being alive.
+  if (process.env.STATE_BUCKET)
+    envelope.state = {
+      getUrl: presign("GET", 3600),
+      putUrl: presign("PUT", 3600),
+      ...(ownUrl && process.env.STATE_REFRESH_SECRET
+        ? {
+            refresh: {
+              url: `${ownUrl.replace(/\/$/, "")}/__fastagent/state-urls`,
+              auth: process.env.STATE_REFRESH_SECRET,
+            },
+          }
+        : {}),
+    };
+  const res = await client.send(
+    new InvokeAgentRuntimeCommand({
+      agentRuntimeArn: process.env.RUNTIME_ARN,
+      runtimeSessionId: process.env.INGRESS_SESSION_ID,
+      contentType: "application/json",
+      accept: "application/json",
+      payload: new TextEncoder().encode(JSON.stringify(envelope)),
+    }),
+  );
+  const body = Buffer.from(await res.response.transformToByteArray());
+  return { status: res.statusCode ?? 200, body };
+}
+
+// Mirror the container's pending wake-ups into one-shot schedules: at(fireAt), poke me, delete
+// after firing. Upsert (create → conflict → update). The container pre-filters DUE alarms (it is
+// awake handling those), so every failure here is REAL — counted and propagated: a swallowed error
+// would leave a pending wake with no alarm, exactly the reliability hole this mechanism closes.
+// Cancelled wakes are NOT deleted here: their poke fires, finds nothing due, and the schedule
+// self-deletes (lazy cleanup by design).
+async function syncAlarms(alarms, ctx) {
+  const { SchedulerClient, CreateScheduleCommand, UpdateScheduleCommand } = require("@aws-sdk/client-scheduler");
+  const sch = new SchedulerClient({});
+  let failed = 0;
+  // Alarm name = a stable hash of the WHOLE wake id. A prefix of the id would collide (two wakes
+  // sharing 8 hex chars), and a collision is INDISTINGUISHABLE from the legitimate re-arm below:
+  // the second wake would "update" the first's alarm and silently steal its fire time.
+  const names = new Map();
+  for (const a of alarms) {
+    const name = process.env.WAKE_PREFIX + crypto.createHash("sha256").update(a.id).digest("hex").slice(0, 16);
+    if (names.has(name)) {
+      failed += 1;
+      console.log(`alarm name collision ${name}: ${names.get(name)} vs ${a.id}`);
+      continue;
+    }
+    names.set(name, a.id);
+    const p = {
+      Name: name,
+      ScheduleExpression: `at(${a.at.slice(0, 19)})`,
+      ScheduleExpressionTimezone: "UTC",
+      FlexibleTimeWindow: { Mode: "OFF" },
+      ActionAfterCompletion: "DELETE",
+      Target: { Arn: ctx.invokedFunctionArn, RoleArn: process.env.WAKE_ROLE_ARN, Input: '{"wakePoke":true}' },
+    };
+    try {
+      await sch.send(new CreateScheduleCommand(p));
+    } catch (e) {
+      try {
+        if (e.name !== "ConflictException") throw e;
+        await sch.send(new UpdateScheduleCommand(p));
+      } catch (u) {
+        failed += 1;
+        console.log(`alarm ${p.Name}: ${u}`);
+      }
+    }
+  }
+  return failed;
+}
+
+exports.handler = async (event, ctx) => {
+  // EventBridge wake-up poke: the invocation itself wakes the container; its pump does the rest.
+  if (event?.wakePoke) {
+    const r = await invoke({ kind: "wake-poke" });
+    console.log(`wake-poke: ${r.status}`);
+    return { status: r.status };
+  }
+  // EventBridge Scheduler fire — throw on failure so the miss lands in CloudWatch, never silently.
+  if (event?.scheduleFire) {
+    const { name, slot } = event.scheduleFire;
+    const r = await invoke({ kind: "schedule-fire", name, slot });
+    const out = r.body.toString();
+    console.log(`schedule-fire ${name} (${slot}): ${r.status} ${out}`);
+    if (r.status >= 400) throw new Error(`schedule-fire ${name} failed: ${r.status} ${out}`);
+    return { status: r.status };
+  }
+  const http = event?.requestContext?.http;
+  if (!http) throw new Error("unrecognized event shape");
+  // Refresh the snapshot capabilities with THIS Lambda invocation's current temporary credentials.
+  // The container may settle long after the webhook Lambda (and its credentials) expired.
+  if (event.rawPath === "/__fastagent/state-urls") {
+    const req = JSON.parse(event.isBase64Encoded ? Buffer.from(event.body, "base64").toString() : event.body || "{}");
+    if (!secretEq(req.auth, process.env.STATE_REFRESH_SECRET)) return { statusCode: 403, body: "forbidden\n" };
+    return {
+      statusCode: 200,
+      headers: { "content-type": "application/json" },
+      body: JSON.stringify({ getUrl: presign("GET", 3600), putUrl: presign("PUT", 3600) }),
+    };
+  }
+  // The container's wake-alarm callback (reserved path, shared secret) — handled HERE, never forwarded.
+  if (event.rawPath === "/__fastagent/wake-alarm") {
+    const req = JSON.parse(event.isBase64Encoded ? Buffer.from(event.body, "base64").toString() : event.body || "{}");
+    if (!secretEq(req.secret, process.env.WAKE_SECRET)) return { statusCode: 403, body: "forbidden\n" };
+    const failed = await syncAlarms(req.alarms || [], ctx);
+    // Partial failure IS failure: the container retries the whole (idempotent) set until every
+    // pending wake really has its alarm.
+    if (failed > 0) return { statusCode: 500, body: `${failed} alarm(s) failed\n` };
+    return { statusCode: 200, body: "ok\n" };
+  }
+  // The deploy driver's probe (reserved path, ingress secret): wake the runtime through the SAME
+  // trusted envelope pipeline (state URLs included — a direct InvokeAgentRuntime call could not mint
+  // them, and would make the runtime construct against a pre-restore mount) and pass its structured
+  // transport-200 verdict back VERBATIM. The ordinary webhook path below folds a non-200 transport
+  // into an opaque 502, which would strip exactly the diagnostics the probe exists to carry — and it
+  // sits BEFORE the WEBHOOKS_ENABLED gate so schedule-only topologies (whose URLs refuse ordinary
+  // public traffic) are probeable too.
+  if (event.rawPath === "/__fastagent/probe") {
+    const req = JSON.parse(event.isBase64Encoded ? Buffer.from(event.body, "base64").toString() : event.body || "{}");
+    if (!secretEq(req.auth, process.env.INGRESS_SECRET)) return { statusCode: 403, body: "forbidden\n" };
+    const r = await invoke({ kind: "probe" });
+    if (r.status !== 200) {
+      console.log(`probe transport error ${r.status}: ${r.body}`);
+      return { statusCode: 502, body: "upstream error\n" };
+    }
+    return { statusCode: 200, headers: { "content-type": "application/json" }, body: r.body.toString() };
+  }
+  // Enforce the advertised ORIGINAL-body ceiling before base64 adds another 4/3 inside the runtime
+  // envelope. This also leaves deterministic room for headers/query/JSON under Lambda's 6 MB cap.
+  const webhookBytes =
+    event.body === undefined
+      ? 0
+      : event.isBase64Encoded
+        ? Buffer.byteLength(event.body, "base64")
+        : Buffer.byteLength(event.body);
+  if (webhookBytes > Number(process.env.MAX_WEBHOOK_BODY_BYTES))
+    return { statusCode: 413, body: "payload too large\n" };
+  // A schedule-only deployment has a Function URL solely for the authenticated refresh callback.
+  // Reject arbitrary public traffic BEFORE it can wake AgentCore (cost/DoS) or reach an inner route.
+  if (process.env.WEBHOOKS_ENABLED !== "1") return { statusCode: 404, body: "not found\n" };
+  // Function URL webhook — forward the original request verbatim (signature material included).
+  const r = await invoke({
+    kind: "webhook",
+    method: http.method,
+    path: event.rawPath || "/",
+    query: event.rawQueryString || undefined,
+    headers: event.headers || {},
+    bodyB64:
+      event.body === undefined
+        ? undefined
+        : event.isBase64Encoded
+          ? event.body
+          : Buffer.from(event.body).toString("base64"),
+  });
+  if (r.status !== 200) {
+    console.log(`transport error ${r.status}: ${r.body}`);
+    return { statusCode: 502, body: "upstream error\n" };
+  }
+  const reply = JSON.parse(r.body.toString()); // { status, headers, bodyB64 } from the adapter
+  for (const k of Object.keys(reply.headers)) {
+    if (/^(content-length|transfer-encoding|connection)$/i.test(k)) delete reply.headers[k];
+  }
+  return { statusCode: reply.status, headers: reply.headers, body: reply.bodyB64, isBase64Encoded: true };
+};
